Global exceptions implemented

+++ src/main/java/com/amazingcode/in/example/service/impl/CategoryServiceImpl.java
@@ -1,10 +1,13 @@
 package com.amazingcode.in.example.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
 import com.amazingcode.in.example.entity.Category;
+import com.amazingcode.in.example.exception.AlreadyPresentException;
+import com.amazingcode.in.example.exception.NotPresentException;
 import com.amazingcode.in.example.repository.CategoryRepository;
 import com.amazingcode.in.example.service.CategoryService;
 
@@ -19,24 +22,36 @@ public class CategoryServiceImpl implements CategoryService {
 
 	@Override
 	public Category createCategory(Category category) {
+		boolean isCategoryPresent = categoryRepository.existsByCategoryName(category.getCategoryName());
+		if(isCategoryPresent) {
+			throw new AlreadyPresentException("Category with name "+category.getCategoryName()+" is already present.");
+		}
 		return categoryRepository.save(category);
 	}
 
 	@Override
 	public List<Category> getAllCategory() {
-		return categoryRepository.findAll();
+		List<Category> existsCategories = categoryRepository.findAll();
+		if(existsCategories.isEmpty()||existsCategories==null) {
+			throw new NotPresentException("Categories not present.");
+		}
+		return existsCategories;
 	}
 
 	@Override
 	public Category getCategory(Long id) {
-		return categoryRepository.findById(id).get();
+		Optional<Category> existCategory = categoryRepository.findById(id);
+		if(existCategory.isEmpty()) {
+			throw new NotPresentException("Category with id "+id+" does not exist.");
+		}
+		return existCategory.get();
 	}
 
 	@Override
 	public Category updateCategory(Long id, Category category) {
-		Category existCategory = categoryRepository.findById(id).get();
-		if (existCategory == null) {
-			return null;
+		Optional<Category> existCategory = categoryRepository.findById(id);
+		if (existCategory.isEmpty()) {
+			throw new NotPresentException("Category with id "+id+" does not exist.");
 		}
 		category.setCategoryId(id);
 		Category updatedCategory = categoryRepository.save(category);
@@ -45,6 +60,10 @@ public class CategoryServiceImpl implements CategoryService {
 
 	@Override
 	public void deleteCategory(Long id) {
+		Optional<Category> existCategory = categoryRepository.findById(id);
+		if (existCategory.isEmpty()) {
+			throw new NotPresentException("Category with id "+id+" does not exist to delete.");
+		}
 		categoryRepository.deleteById(id);
 	}
